requirements.txt
# Core dependencies for medical document analysis
pandas>=1.5.0
mammoth>=1.4.0
openpyxl>=3.0.0

# AI API clients
openai>=1.0.0
anthropic>=0.7.0
google-generative-ai

# Additional utilities
requests>=2.25.0